refactor: handle speed model failures in ActiveContour
- introduce SpeedModelStatus (init/runtime failures)
- propagate failures from speed model to ActiveContour
- add StoppingStatus::SpeedModelFailure
- stop contour when speed model becomes invalid
- improve robustness of region-based speed computation

+++ src/app/core/frame_clock.cpp
@@ -25,7 +25,7 @@ int64_t FrameClock::nowNs()
 
 double FrameClock::nowSec()
 {
-    return nowNs() * 1e-9;
+    return static_cast<double>(nowNs()) * 1e-9;
 }
 
 } // namespace fluvel_app
